SQL Universe file.
The Celestial Bodies Database project is designed to provide hands-on experience with PostgreSQL, one of the most widely used relational database management systems in the industry. This project focuses on designing, creating, and managing a database that models a universe consisting of various celestial entities such as galaxies, stars, planets, satellites, and moons. The goal is to gain familiarity with database operations, SQL queries, and data management, which are crucial skills for any data-related job role.